Try adding to breadcrumbs directly
