+++ corehq/middleware.py
@@ -13,7 +13,6 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib.auth.views import LogoutView
 from django.utils.deprecation import MiddlewareMixin
-from sentry_sdk import add_breadcrumb
 
 from corehq.apps.domain.models import Domain
 from corehq.apps.domain.utils import legacy_domain_re
@@ -120,17 +119,7 @@ class LogLongRequestMiddleware(MiddlewareMixin):
     def process_response(self, request, response):
         request_timer = getattr(response, 'request_timer', None)
         if request_timer:
-
-            def visit(element, prefix=""):
-                if element.duration is not None and element.percent_of_total is not None:
-                    message = (f"⏱  {element.percent_of_total:>3.0f}% {prefix} "
-                               f"{element.name or ''}: {element.duration:0.3f}s")
-                    add_breadcrumb(category="timing", message=message, level="info")
-                prefix += "  →"
-                for sub in element.subs:
-                    visit(sub, prefix=prefix)
-
-            visit(request_timer.root)
+            request_timer.add_to_sentry_breadcrumbs()
 
         if hasattr(request, '_profile_starttime'):
             duration = datetime.datetime.utcnow() - request._profile_starttime
